correct dependencies
