+++ ydb/core/tx/columnshard/engines/scheme/indexes/ya.make
@@ -2,7 +2,6 @@ LIBRARY()
 
 PEERDIR(
     ydb/core/tx/columnshard/engines/scheme/indexes/abstract
-    ydb/core/tx/columnshard/engines/scheme/indexes/bloom
 )
 
 END()
