+++ ydb/core/tx/columnshard/engines/storage/indexes/bloom/ya.make
@@ -9,6 +9,7 @@ SRCS(
 PEERDIR(
     ydb/core/protos
     ydb/core/formats/arrow
+    ydb/core/tx/columnshard/engines/storage/indexes/portions
 )
 
 END()
