+++ ydb/core/tx/columnshard/engines/storage/indexes/ya.make
@@ -1,0 +1,8 @@
+LIBRARY()
+
+PEERDIR(
+    ydb/core/tx/columnshard/engines/storage/indexes/portions
+    ydb/core/tx/columnshard/engines/storage/indexes/bloom
+)
+
+END()
