+++ ydb/core/tx/columnshard/engines/storage/optimizer/abstract/optimizer.h
@@ -186,6 +186,12 @@ private:
 
 public:
 
+    static std::shared_ptr<IOptimizerPlannerConstructor> BuildDefault() {
+        auto result = TFactory::MakeHolder("l-buckets");
+        AFL_VERIFY(!!result);
+        return std::shared_ptr<IOptimizerPlannerConstructor>(result.Release());
+    }
+
     virtual ~IOptimizerPlannerConstructor() = default;
 
     bool ApplyToCurrentObject(const std::shared_ptr<IOptimizerPlanner>& current) const {
