+++ ydb/core/tx/schemeshard/olap/indexes/ya.make
@@ -7,6 +7,7 @@ SRCS(
 
 PEERDIR(
     ydb/services/bg_tasks/abstract
+    ydb/core/tx/columnshard/engines/scheme/indexes/abstract
 )
 
 END()
