+++ ydb/core/tx/schemeshard/olap/operations/alter/in_store/transfer/ya.make
@@ -6,6 +6,7 @@ SRCS(
 
 PEERDIR(
     ydb/core/tx/schemeshard/olap/operations/alter/abstract
+    ydb/core/tx/columnshard/data_sharing/initiator/controller
     ydb/core/tx/columnshard/data_sharing/protos
 )
